Исправления замечаний, полученных в ходе ревью.
