+++ service/src/main/java/ru/practicum/service/controller/events/PublicEventsController.java
@@ -3,7 +3,6 @@ package ru.practicum.service.controller.events;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.constraints.PositiveOrZero;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -24,7 +23,6 @@ public class PublicEventsController {
     private final StatisticsService statisticsService;
     private final EventsService eventsService;
 
-    @Autowired
     public PublicEventsController(StatisticsService statisticsService,
                                   EventsService eventsService) {
         this.statisticsService = statisticsService;
