+++ service/src/main/java/ru/practicum/service/controller/requests/RequestsController.java
@@ -1,7 +1,6 @@
 package ru.practicum.service.controller.requests;
 
 import jakarta.validation.constraints.PositiveOrZero;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -16,7 +15,6 @@ import java.util.List;
 public class RequestsController {
     RequestsService requestsService;
 
-    @Autowired
     public RequestsController(RequestsService requestsService) {
         this.requestsService = requestsService;
     }
