+++ service/src/main/java/ru/practicum/service/controller/user/UserController.java
@@ -1,7 +1,6 @@
 package ru.practicum.service.controller.user;
 
 import jakarta.validation.constraints.Min;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -17,7 +16,6 @@ import java.util.List;
 public class UserController {
     UserService userService;
 
-    @Autowired
     public UserController(UserService userService) {
         this.userService = userService;
     }
